requirements.txt
click==7.1.2
numpy==1.20.1
pandas==1.2.3
kfp==1.4.0
scikit-learn==0.24.1
seldon-core==1.6.0
boto3==1.17.22
spacy==3.0.3
dill==0.3.3
kubeflow-metadata==0.3.1
kfp_server_api==1.4.1